crud final

+++ Nutrin/Consulta/Services/Horarios/deleteHorario.py
@@ -3,10 +3,14 @@ from Nutrin.Consulta.Services.Horarios.readHorario import readHorario
 
 def deleteHorario(data, hora):
     status, dado = readHorario(data, hora, True)
+    
     if status:
-        db.session.delete(dado)
-        db.session.commit()
-        return True, "Horário excluido"
+        if dado.utilizado:
+            return False, "Não é possivel excluir. Há um paciente utilizando o horário"
+        else:
+            db.session.delete(dado)
+            db.session.commit()
+            return True, "Horário excluido"
     return status, dado
